requirements.txt
beautifulsoup4
requests
selenium
webdriver-manager
youtube-comment-downloader
Pillow
Image
colormath
counter
io
bytesIO
json
LabColor
numpy
urllib